Split IsExperimentalFeatureEnabled to new file
