+++ confgenerator/feature_tracking.go
@@ -3,7 +3,6 @@ package confgenerator
 import (
 	"errors"
 	"fmt"
-	"os"
 	"reflect"
 	"sort"
 	"strings"
@@ -272,14 +271,4 @@ func getOverriddenDefaultPipelines(uc *UnifiedConfig) []Feature {
 	}
 
 	return features
-}
-
-func IsExperimentalFeatureEnabled(receiver string) bool {
-	enabledList := strings.Split(os.Getenv("EXPERIMENTAL_FEATURES"), ",")
-	for _, e := range enabledList {
-		if e == receiver {
-			return true
-		}
-	}
-	return false
 }
